requirements.txt
requests
python-dotenv
ollama
chromadb
pydantic
pandas
langchain-community
langchain-core
langchain-ollama
torch
streamlit
pypdf